Add results_logger for logging per-second results ...
And config stuff for it. The intention is to log to a single static file
in python and for the operator to use logrotate if they don't like that.
This file will grow quite large over time, and while python has a
RotatingFileHandler [0], it doesn't have a way to compress when
rotating [1] like logrotate does. Thus the WatchedFileHandler [0],
which is made to support external utils like logrotate moving the file
out from under python.

[0]: https://docs.python.org/3/library/logging.handlers.html
[1]: https://stackoverflow.com/a/8468041